+++ apps/shared/avifutil.c
@@ -10,6 +10,7 @@ void avifImageDump(avifImage * avif)
     printf(" * Resolution   : %dx%d\n", avif->width, avif->height);
     printf(" * Bit Depth    : %d\n", avif->depth);
     printf(" * Format       : %s\n", avifPixelFormatToString(avif->yuvFormat));
+    printf(" * Alpha        : %s\n", (avif->alphaPlane && (avif->alphaRowBytes > 0)) ? "Present" : "Absent");
     switch (avif->profileFormat) {
         case AVIF_PROFILE_FORMAT_NONE:
             printf(" * Color Profile: None\n");