+++ src/codec_aom.c
@@ -304,12 +304,10 @@ static avifBool aomCodecEncodeImage(avifCodec * codec, avifImage * image, avifEn
     int minQuantizer = AVIF_CLAMP(encoder->minQuantizer, 0, 63);
     int maxQuantizer = AVIF_CLAMP(encoder->maxQuantizer, 0, 63);
     if (alpha) {
-        minQuantizer = AVIF_QUANTIZER_LOSSLESS;
-        maxQuantizer = AVIF_QUANTIZER_LOSSLESS;
+        minQuantizer = AVIF_CLAMP(encoder->minQuantizerAlpha, 0, 63);
+        maxQuantizer = AVIF_CLAMP(encoder->maxQuantizerAlpha, 0, 63);
     }
-    avifBool lossless = ((encoder->minQuantizer == AVIF_QUANTIZER_LOSSLESS) && (encoder->maxQuantizer == AVIF_QUANTIZER_LOSSLESS))
-                            ? AVIF_TRUE
-                            : AVIF_FALSE;
+    avifBool lossless = ((minQuantizer == AVIF_QUANTIZER_LOSSLESS) && (maxQuantizer == AVIF_QUANTIZER_LOSSLESS)) ? AVIF_TRUE : AVIF_FALSE;
     cfg.rc_min_quantizer = minQuantizer;
     cfg.rc_max_quantizer = maxQuantizer;